requirements.txt
influxdb
requests
PyYAML